requirements.txt
functions-framework==3.*
google-cloud-storage
scikit-surprise==1.1.3
pandas
numpy
